fix: correctness and robustness improvements across core/solvers/utils
- core/models_ms.py: fix DeepONetMS FNNLayer depth (branch_depth-2) to match
  DDESolver layer count; fix HEAQNNMS/FNNMS net_size unpack to tolerate 4-tuple
- core/models_pt.py: remove spurious bias parameter from HEAQNNPT to match
  original HEAQNNMS (no bias)
- compare_backends.py: align DeepONet_PT inline layer count and drop bias key
  from HEAQNN test state dict
- solvers/solver_ms.py, solver_pt.py: switch Rel_L2 from per-batch average to
  global sqrt(SSE)/sqrt(norm²); fix num_batches to use ceil division
- solvers/solver_dde.py: unify LR scheduler key (step_size / decay_steps) so
  shared configs work across all three solvers
- infer.py: extend _TF_RE regex to match _FF_ checkpoints (non-TF models)
- utils/common.py: fix --ham_bound argument type int→float
- utils/weight_transfer.py: fix load_quanonet_pt scale_coeff default (0.01);
  add if_trainable_freq param to ms_npz_to_pt_state_dict to skip LinearLayer
  key lookup for non-TF checkpoints
- ibm_inference.py: fix min_num_qubits to use actual n_qubits instead of 127

+++ utils/common.py
@@ -76,7 +76,7 @@ def get_base_parser():
     parser.add_argument('--num_qubits', type=int, default=None, help='[Quantum] Number of qubits (default: 5)')
     parser.add_argument('--scale_coeff', type=float, help='[Quantum] Scaling coefficient')
     parser.add_argument('--if_trainable_freq', type=str, default=None, help='[Quantum] Trainable frequency (default: true)')
-    parser.add_argument('--ham_bound', type=int, nargs='+', default=None,
+    parser.add_argument('--ham_bound', type=float, nargs='+', default=None,
                     help='[Quantum] Hamiltonian bounds (e.g., --ham_bound -5 5) (default: -5 5)')
     parser.add_argument('--ham_pauli', type=str, default=None, choices=['X', 'Y', 'Z'],
                         help='Pauli observable basis for the Hamiltonian (default: Z).')
